rm useless funcs and files
