+++ b.cub
@@ -12,5 +12,5 @@ C 90,90,80
 1000000000000000000000001    11111
 1000000000111110000000001
 1000000001     1000000001    111111
-100000001       1000N0001    100001
+100000001       1000E0001    100001
 11111111         11111111    11111
