+++ src/game/gameplay.c
@@ -27,9 +27,6 @@ void screen(t_game *game)
 		exit(1);
 	}
     raycast(game);
-    //for debugging or minimap..
-    //draw_map(game);
-    //draw_player(game);
 }
 
 void keys(void  *param)
@@ -44,61 +41,34 @@ void keys(void  *param)
     screen(game);
 }
 
-// void init_game(t_map *map_info, t_game *game, t_player *player, t_txtr *txtr, t_ray *r)
-// {
-//     if (!(game->mlx = mlx_init(WIN_WIDTH, WIN_HEIGHT, WIN_NAME, true)))
-// 		exit(1);
-//     init_player(map_info, player);
-//     game->player = player;
-//     load_textures(txtr);
-//     game->ray = r;
-// }
-
-// int gameplay(t_map *map_info, t_txtr *txtr)
-// {
-//     t_game game;
-//     t_player player;
-//     t_ray r;
-
-//     game = (t_game){0};
-//     player = (t_player){0};
-//     r = (t_ray){0};
-//     game.textures = txtr;
-//     game.map = map_info->temp_map;
-//     init_game(map_info, &game, &player, txtr, &r);
-//     screen(&game);
-//     mlx_loop_hook(game.mlx, keys, &game);
-// 	mlx_loop(game.mlx);
-// 	mlx_terminate(game.mlx);
-//     free_test_map(&game);
-//     return (0);
-// }
-
-void init_game(t_game *game)
+void load_textures(t_txtr *txtr)
 {
-    if (!(game->mlx = mlx_init(WIN_WIDTH, WIN_HEIGHT, WIN_NAME, true)))
-		exit(1);
-    // init_player(game);
-    // game->player = player;
-    load_textures(game->textures);
-    // printf("%f %f %f\n", game->player->ppx, game->player->ppy, game->player->pa);
-    // game->ray = r;
+    txtr->n_txtr = mlx_load_png(txtr->info[0]);
+    if (!txtr->n_txtr)
+        exit(12);
+    txtr->s_txtr = mlx_load_png(txtr->info[1]);
+    if (!txtr->s_txtr)
+        exit(12);
+    txtr->w_txtr = mlx_load_png(txtr->info[2]);
+    if (!txtr->w_txtr)
+        exit(12);
+    txtr->e_txtr = mlx_load_png(txtr->info[3]);
+    if (!txtr->e_txtr)
+        exit(12);
 }
 
 int gameplay(t_game *game)
 {
     t_ray       r;
-    // t_player    p;
 
     r = (t_ray){0};
-    // p = (t_player){0};
     game->ray = &r;
-    // game->player = &p;
-    init_game(game);
+    if (!(game->mlx = mlx_init(WIN_WIDTH, WIN_HEIGHT, WIN_NAME, true)))
+        exit(1);
+    load_textures(game->textures);
     screen(game);
     mlx_loop_hook(game->mlx, keys, game);
 	mlx_loop(game->mlx);
-	mlx_terminate(game->mlx);
-    // free_test_map(&game);
+	mlx_terminate(game->mlx);/*tilalle exit loop joka putsaa structit ja sulkee pelin?*/
     return (0);
 }
