+++ src/parse/validate.c
@@ -23,16 +23,16 @@
 
 static int	compare_rows(size_t shorter, size_t longer, char *long_row)
 {
-	size_t	i;
+	size_t	x;
 
 	if (!long_row)
 		return (1);
-	i = shorter;
-	while (i < longer)
+	x = shorter;
+	while (x < longer)
 	{
-		if (long_row[i] != '1')
+		if (long_row[x] != '1')
 			return (1);
-		i++;
+		x++;
 	}
 	return (0);
 }
@@ -42,28 +42,28 @@ static int	compare_rows(size_t shorter, size_t longer, char *long_row)
  */
 static int	wall_coverage(t_map *map_info)
 {
-	size_t	k;
+	size_t	y;
 	size_t	below;
 	size_t	current;
 	size_t	above;
 
-	k = 1;
-	while (k < (map_info->size_y - 1))
+	y = 1;
+	while (y < (map_info->size_y - 1))
 	{
-		below = ft_strlen(map_info->temp_map[k - 1]);
-		current = ft_strlen(map_info->temp_map[k]);
-		above = ft_strlen(map_info->temp_map[k + 1]);
-		if (!map_info->temp_map[k] || !map_info->temp_map[k - 1] || !map_info->temp_map[k + 1])
+		below = ft_strlen(map_info->temp_map[y - 1]);
+		current = ft_strlen(map_info->temp_map[y]);
+		above = ft_strlen(map_info->temp_map[y + 1]);
+		if (!map_info->temp_map[y] || !map_info->temp_map[y - 1] || !map_info->temp_map[y + 1])
 			return (1);
-		if (current > below && compare_rows(below, current, map_info->temp_map[k]))
+		if (current > below && compare_rows(below, current, map_info->temp_map[y]))
 			return (1);
-		else if (current < below && compare_rows(current, below, map_info->temp_map[k - 1]))
+		else if (current < below && compare_rows(current, below, map_info->temp_map[y - 1]))
 			return (1);
-		if (current > above && compare_rows(above, current, map_info->temp_map[k]))
+		if (current > above && compare_rows(above, current, map_info->temp_map[y]))
 			return (1);
-		else if (current < above && compare_rows(current, above, map_info->temp_map[k + 1]))
+		else if (current < above && compare_rows(current, above, map_info->temp_map[y + 1]))
 			return (1);
-		k++;
+		y++;
 	}
 	return (0);
 }
@@ -74,23 +74,23 @@ static int	wall_coverage(t_map *map_info)
  */
 static int	first_and_last_row(t_map *map_info)
 {
-	int	i;
-	int	j;
+	int	x;
+	int	y;
 
-	i = 0;
-	while (map_info->temp_map[0][i])
+	x = 0;
+	while (map_info->temp_map[0][x])
 	{
-		if (map_info->temp_map[0][i] != '1')
+		if (map_info->temp_map[0][x] != '1')
 			return (write_err("invalid first row"), 1);
-		i++;
+		x++;
 	}
-	i = 0;
-	j = map_info->size_y - 1;
-	while (map_info->temp_map[j][i])
+	x = 0;
+	y = map_info->size_y - 1;
+	while (map_info->temp_map[y][x])
 	{
-		if (map_info->temp_map[j][i] != '1')
+		if (map_info->temp_map[y][x] != '1')
 			return (write_err("invalid last row"), 1);
-		i++;
+		x++;
 	}
 	if (wall_coverage(map_info))
 		return (write_err("invalid map walls"), 1);
@@ -105,33 +105,28 @@ static int	first_and_last_row(t_map *map_info)
 
 static int validate_chars(char **s, t_game *game)
 {
-    size_t		i;
-	size_t		k;
+    size_t		x;
+	size_t		y;
 
-	k = -1;
-	while (s[++k])
+	y = -1;
+	while (s[++y])
 	{
-		i = -1;
-		while (s[k][++i])
+		x = -1;
+		while (s[y][++x])
 		{
-			if (!is_player(s[k][i]) && s[k][i] != ' ' && s[k][i] != '0' && s[k][i] != '1')
+			if (!is_player(s[y][x]) && s[y][x] != ' ' && s[y][x] != '0' && s[y][x] != '1')
 				return (1);
-			if (is_player(s[k][i]))
-			{
-				game->player->ppx = (float)i * 64;
-    			game->player->ppy = (float)k * 64;
-				game->player->pa = PI / 2;/*add func for start orientation*/
-			}
-			if (s[k][i] == ' ')
-				s[k][i] = '1';
+			if (is_player(s[y][x]))
+				set_player(game, s[y][x], x, y);
+			if (s[y][x] == ' ')
+				s[y][x] = '1';
 		}
-		if (s[k][i - 1] != '1')
+		if (s[y][x - 1] != '1')
 			return (write_err("incorrect map"), 1);
 		game->map_info->size_y++;
 	}
 	if (first_and_last_row(game->map_info))
 		return (1);
-	printf("%f %f %f\n", game->player->ppx, game->player->ppy, game->player->pa);
 	return (0);
 }
 
@@ -152,11 +147,11 @@ int	process_map(t_game *game)
 	if (get_info(game))
 	{
 		free_arr(game->textures->info);
-        // free(txtr);
         return (1);
 	}
 	if (validate_chars(game->map_info->temp_map, game))
 		return (free_arr(game->map_info->temp_map), free_arr(game->textures->info), 1);
+	// printf("%f %f %f\n", game->player->ppx, game->player->ppy, game->player->pa);
 	if (gameplay(game))
 		return (free_arr(game->textures->info), 1);
 	return (0);
