--- src/render/test_map.c
@@ -1,76 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   test_map.c                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: eleppala <eleppala@student.hive.fi>        +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/10/17 15:43:13 by eleppala          #+#    #+#             */
-/*   Updated: 2024/10/17 15:43:15 by eleppala         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../../incs/cub3D.h"
-
-// void test_map(t_game *game)
-// {
-//     int i;
-//     int j;
-
-//     i = 0;
-//     j = 0;
-//     game->map = ft_split("1111111111-"
-//                         "1000000001-"
-//                         "1000111111-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "10001-"
-//                         "11111", '-');
-
-
-//     //game->map = ft_split("111111111111111-100100000000001-100100000000001-100100010000001-100000000000001-100000000000001-100000000000001-100000000000001-100000000000001-111111111111111", '-');
-//     //window->map = ft_split("1111111111-1000100001-1000000001-1000000001-1000000001-1000000001-1000000001-1000000001-1000000001-1111111111", '-');
-//     //window->map = ft_split("11111-10001-10001-10001-11111", '-');
-//     if (!game->map)
-//         exit(1);
-//     while (game->map[i])
-//     {
-//         while(game->map[i][j])
-//         {
-//             j ++;
-//         }
-//         if (j > game->map_width)
-//             game->map_width = j;
-//         j = 0;
-//         i ++;
-//     }
-//     game->map_width = game->map_width * TILE;
-//     game->map_height = i * TILE;
-// }
-
-void free_test_map(t_game *game)
-{
-    size_t i = 0;
-
-    while(game->map_info->temp_map[i] != NULL)
-    {
-        free(game->map_info->temp_map[i]);
-        i++;
-    }
-    free(game->map_info->temp_map);
-    game->map_info->temp_map = NULL;
-}
